\name{MExPosition-package}
\alias{MExPosition-package}
\alias{MExPosition}
\encoding{UTF-8} 
\docType{package}
\title{\emph{M}ulti-table \emph{Ex}ploratory Analysis with the Singular Value Decom\emph{Position} with the STATIS family.}
\description{
MExPosition is multi-table \code{\link{ExPosition}} and includes the family of STATIS method, such as Plain STATIS, DISTATIS, Dual STATIS and ANISOSTATIS. The core of MExPosition is  \code{\link{ExPosition}} and the \code{\link{svd}}.
}
\details{
\tabular{ll}{
Package: \tab MExPosition\cr
Type: \tab Package\cr
Version: \tab 2.0.3\cr
Date: \tab 2013-06-10\cr
Depends: \tab R (>=2.15.0), prettyGraphs (>= 2.0.0), ExPosition (>= 2.0.0)\cr
License: \tab GPL-2\cr
}
}
\author{
Questions, comments, compliments, and complaints go to Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}.\cr\cr

The following people are contributors to MExPosition code, data, or examples:\cr
Derek Beaton and Hervé Abdi.
}
\references{
Abdi, H., Valentin, D., Chollet, S., & Chrea, C. (2007). Analyzing assessors and products in sorting tasks: DISTATIS, theory and applications. \emph{Food Quality and Preference}, 18, 627-640.\cr\cr
Abdi, H., & Valentin, D. (2005). DISTATIS: the analysis of multiple distance matrices. In N.J. Salkind (Ed.): \emph{Encyclopedia of Measurement and Statistics}. Thousand Oaks (CA): Sage. pp. 284-290.\cr\cr
Abdi, H., Williams, L.J., Valentin, D., & Bennani-Dosse, M. (2012). STATIS and DISTATIS: Optimum multi-table principal component analysis and three way metric multidimensional scaling. \emph{Wiley Interdisciplinary Reviews: Computational Statistics}, 4. \cr\cr
Abdi, H., & Valentin, D. (2007). STATIS. In N.J. Salkind (Ed.): \emph{Encyclopedia of Measurement and Statistics}. Sage. pp. 955-962.\cr
}
\keyword{package}
\keyword{multivariate}
\seealso{\code{\link{mpSTATIS}}, \code{\link{mpDISTATIS}}
}
\examples{
#For more examples, see each individual function (as noted above).
}
