\name{print.covstatis.innerproduct}
\alias{print.covstatis.innerproduct}
\title{Print the results of the Inner Product for COVSTATIS}
\description{S3 Class function to print results for MExPosition.}
\usage{\method{print}{covstatis.innerproduct}(x,\dots)}
\arguments{
  \item{x}{an object of class COVSTATIS}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}