\name{print.covstatis.overview}
\alias{print.covstatis.overview}
\title{Print the Overview for COVSTATIS}
\description{S3 Class function to print results for MExPosition.}
\usage{\method{print}{covstatis.overview}(x,\dots)}
\arguments{
  \item{x}{an object of class COVSTATIS}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}