\name{print.distatis.overview}
\alias{print.distatis.overview}
\title{Print the results of the Overview of DISTATIS}
\description{S3 Class function to print the results for MExPosition.}
\usage{\method{print}{distatis.overview}(x,\dots)}
\arguments{
  \item{x}{an object of class STATIS}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt <cherise.chinfatt@utdallas.edu>}
