\name{print.doact.statis.compromise}
\alias{print.doact.statis.compromise}
\title{Print the results of the Compromise for DO-ACT}
\description{S3 Class function to print results for MExPosition.}
\usage{\method{print}{doact.statis.compromise}(x,\dots)}
\arguments{
  \item{x}{an object of class DO-ACT}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}