\name{print.doact.statis.innerproduct}
\alias{print.doact.statis.innerproduct}
\title{Print the results of the Inner Product for DO-ACT}
\description{S3 Class function to print results for MExPosition.}
\usage{\method{print}{doact.statis.innerproduct}(x,\dots)}
\arguments{
  \item{x}{an object of class Do-ACT}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}