\name{print.KPlus1.statis.compromise}
\alias{print.KPlus1.statis.compromise}
\title{Print the results of the Compromise for (K+1) STATIS}
\description{S3 Class function to print results for MExPosition.}
\usage{\method{print}{KPlus1.statis.compromise}(x,\dots)}
\arguments{
  \item{x}{an object of class KPlus1.statis}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}