\name{print.KPlus1.statis.innerproduct}
\alias{print.KPlus1.statis.innerproduct}
\title{Print the results of the Inner Product for (K+1) STATIS}
\description{S3 Class function to print results for MExPosition.}
\usage{\method{print}{KPlus1.statis.innerproduct}(x,\dots)}
\arguments{
  \item{x}{an object of class KPlus1.statis}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}