\name{print.mexPosition.Output}
\alias{print.mexPosition.Output}
\title{Print the results of MExPosition}
\description{S3 Class function to print the results for MExPosition.}
\usage{\method{print}{mexPosition.Output}(x,\dots)}
\arguments{
  \item{x}{an object of class STATIS}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt <cherise.chinfatt@utdallas.edu>}
