\name{print.mpANISOSTATIS}
\alias{print.mpANISOSTATIS}
\title{Print ANISOSTATIS results}
\description{S3 Class function to print ANISOSTATIS results.}
\usage{\method{print}{mpANISOSTATIS}(x,\dots)}
\arguments{
  \item{x}{list that contains items to make into the mpANISOSTATIS class.}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}