\name{print.mpCOVSTATIS}
\alias{print.mpCOVSTATIS}
\title{Print COVSTATIS results}
\description{S3 Class function to print COVSTATIS results.}
\usage{\method{print}{mpCOVSTATIS}(x,\dots)}
\arguments{
  \item{x}{list that contains items to make into the mpCOVSTATIS class.}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}