\name{print.mpDISTATIS}
\alias{print.mpDISTATIS}
\title{Print DISTATIS results}
\description{S3 Class function to print DISTATIS results.}
\usage{\method{print}{mpDISTATIS}(x,\dots)}
\arguments{
  \item{x}{list that contains items to make into the mpDISTATIS class.}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}