\name{print.mpDOACT.STATIS}
\alias{print.mpDOACT.STATIS}
\title{Print DOACT.STATIS results}
\description{S3 Class function to print DOACT.STATIS results.}
\usage{\method{print}{mpDOACT.STATIS}(x,\dots)}
\arguments{
  \item{x}{list that contains items to make into the mpDOACT.STATIS class.}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}