\name{print.mpKPlus1STATIS}
\alias{print.mpKPlus1STATIS}
\title{Print KPlus1STATIS results}
\description{S3 Class function to print KPlus1STATIS results.}
\usage{\method{print}{mpKPlus1STATIS}(x,\dots)}
\arguments{
  \item{x}{list that contains items to make into the mpKPlus1STATIS class.}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}