\name{print.mpMFA}
\alias{print.mpMFA}
\title{Print MFA results}
\description{S3 Class function to print the MFA results.}
\usage{\method{print}{mpMFA}(x,\dots)}
\arguments{
  \item{x}{list that contains items to make into the mpMFA class.}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}