\name{print.mpSTATIS}
\alias{print.mpSTATIS}
\title{Print STATIS results}
\description{S3 Class function to print the STATIS results.}
\usage{\method{print}{mpSTATIS}(x,\dots)}
\arguments{
  \item{x}{list that contains items to make into the mpSTATIS class.}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt \email{cherise.chinfatt@utdallas.edu}
}