\name{print.statis.compromise}
\alias{print.statis.compromise}
\title{Print the results for the Compromise of STATIS}
\description{S3 Class function to print the results for MExPosition.}
\usage{\method{print}{statis.compromise}(x,\dots)}
\arguments{
  \item{x}{an object of class STATIS}
 \item{\dots}{inherited/passed arguments for S3 print method(s).}    
}

\author{Cherise R. Chin Fatt <cherise.chinfatt@utdallas.edu>}
